+++ src/Repository/SshKeyRepository.php
@@ -52,18 +52,29 @@ class SshKeyRepository extends ApiRepository
         return new SshKey($sshKeyArray);
     }
 
-    public function create(string $sshKey, string $sshKeyDescription): void
+    public function create(string $sshKey, string $sshKeyDescription, ?bool $isDefault = false): void
     {
         $this->httpClient->post($this->getResourceUrl(), [
             'sshKey'      => $sshKey,
             'description' => $sshKeyDescription,
+            'isDefault'   => $isDefault
         ]);
     }
 
     public function update(int $sshKeyId, string $sshKeyDescription): void
     {
-        $this->httpClient->put($this->getResourceUrl($sshKeyId), ['description' => $sshKeyDescription]);
+        // fetch current ssh key to preserve the current isDefault value. Needed to preserve backwards compatiblity
+        // for the sshkey:setdescription command.
+        $sshKey = $this->getById((string) $sshKeyId);
+        $sshKey->setDescription($sshKeyDescription);
+        $this->updateKey($sshKey);
     }
+
+    public function updateKey(SshKey $sshKey): void
+    {
+        $this->httpClient->put($this->getResourceUrl($sshKey->getId()), ['sshKey' => $sshKey]);
+    }
+
 
     public function delete(int $sshKeyId): void
     {